back

+++ HERErequest.py
@@ -106,21 +106,21 @@ def checkTileFromCache(tile:tuple, layer:str, session:requests.Session=None):
         cache_file_path = f'{tiles_cache_path}{layer}-{tile[2]}-{tile[0]}-{tile[1]}.json'    
 
         ### Add code
-        #if(os.stat(cache_file_path).st_size > 0):
-        #    with open(cache_file_path) as json_file:
-        #        tile_data = json.load(json_file)
-        #else:
-        print("empty layer file")
-        os.remove(cache_file_path)
-        print(f"request data from {layer}")
-        tile_data = getTileRequest(tile, layer, session)['Rows']
-        tileToFile(tile_data, tile, layer)
+        if(os.stat(cache_file_path).st_size > 0):
+            with open(cache_file_path) as json_file:
+                tile_data = json.load(json_file)
+        else:
+            print("empty layer file")
+            os.remove(cache_file_path)
+            print(f"request data from {layer}")
+            tile_data = getTileRequest(tile, layer, session)['Rows']
+            tileToFile(tile_data, tile, layer)
         ### Remove add code
 
         ### Uncoment follow code
         #with open(cache_file_path) as json_file:
         #    tile_data = json.load(json_file)
-        #return tile_data
+        return tile_data
     except:
         print(f"request data from {layer}-{tile[2]}-{tile[0]}-{tile[1]}.json")
         try:
@@ -648,8 +648,8 @@ def setRoadTypes(links_dict,cfg, links_basic_attributes):
     if(cfg['region'] == 'us'):
         #print(links_dict)
         for link_id in links_dict:
-            print(link_id)
-            print(links_dict[link_id]['FUNCTIONAL_CLASS'])            
+            #print(link_id)
+            #print(links_dict[link_id]['FUNCTIONAL_CLASS'])            
             if(links_dict[link_id]['FUNCTIONAL_CLASS'] in [1,2]):
                 links_dict[link_id]['HIGHWAY'] = 'Y'
                 links_dict[link_id]['CITY'] = 'N'
